cleanup proto buffers

+++ BERT/BERT/include/string_utilities.h
@@ -30,9 +30,9 @@ public:
    * icase string comparison, ignoring locale issues (this is for windows paths)
    */
   static int ICaseCompare(const std::string &a, const std::string &b) {
-    int len = a.length();
+    size_t len = a.length();
     if (len != b.length()) return 1;
-    for (int i = 0; i < len; i++) {
+    for (size_t i = 0; i < len; i++) {
       if (toupper(a[i]) != toupper(b[i])) return 2;
     }
     return 0;
